split up collect-css into many helpers
