+++ packages/core/src/generators/react.ts
@@ -9,7 +9,7 @@ import { functionLiteralPrefix } from '../constants/function-literal-prefix';
 import { methodLiteralPrefix } from '../constants/method-literal-prefix';
 import { babelTransformExpression } from '../helpers/babel-transform';
 import { capitalize } from '../helpers/capitalize';
-import { collectCss, collectStyledComponents, hasStyles } from '../helpers/styles/collect';
+import { collectCss } from '../helpers/styles/collect-css';
 import { createMitosisNode } from '../helpers/create-mitosis-node';
 import { fastClone } from '../helpers/fast-clone';
 import { filterEmptyTextNodes } from '../helpers/filter-empty-text-nodes';
@@ -41,6 +41,8 @@ import { MitosisComponent } from '../types/mitosis-component';
 import { MitosisNode } from '../types/mitosis-node';
 import { hasContext } from './helpers/context';
 import { collectReactNativeStyles } from './react-native';
+import { collectStyledComponents } from 'src/helpers/styles/collect-styled-components';
+import { hasStyles } from 'src/helpers/styles/helpers';
 
 export interface ToReactOptions extends BaseTranspilerOptions {
   stylesType?: 'emotion' | 'styled-components' | 'styled-jsx' | 'react-native';
