+++ packages/core/src/generators/template.ts
@@ -1,5 +1,5 @@
 import { format } from 'prettier/standalone';
-import { collectCss } from '../helpers/styles/collect';
+import { collectCss } from '../helpers/styles/collect-css';
 import { fastClone } from '../helpers/fast-clone';
 import { selfClosingTags } from '../parsers/jsx';
 import { MitosisNode } from '../types/mitosis-node';
